UI: Move to builder pattern for compatibility
This fixes builds for Ubuntu 18.04 as they use an older GTK version
than is required (GTK < 3.30) by our codebase. This version does not support Gtk
Templates.

This commit moves the codebase back to using GTK builders. The
original class inheritance structure is preserved so if we want to move back
to the templates method one day this is easy to achieve.

Signed-off-by: jwijenbergh <jeroenwijenbergh@protonmail.com>