UI: Add a connection info expander
This commit adds a connection info expander to the UI.

Uses the NM api for querying, except where I could not find any
functions (e.g. rx and tx bytes, which are gathered from standard file
locations).

If it cannot find the data, as this is not foolproof, it will set it
to N/A, the application will keep working.

Note that glade changed the property names from `_` to `-`, but this
still works on the VMS. Apparently `_` are an old format that is not
used any more in glade.

Signed-off-by: jwijenbergh <jeroenwijenbergh@protonmail.com>