Fix issue #470
During the fix in ccc7319212e2c437d0ae78c8e35f75932b506f42 the `Ovpn.write`
method was rewritten to allow a config that is a list. The config list is
generated by `Ovpn.parse`, which must be used instead of the constructor.

This was tested from commandline with eduvpn-cli configure.