Network: Fix bug when reconnecting with OpenVPN
This fixes a major bug where OpenVPN doesn't work on a reconnect. This
bug is not introduced by this PR, it's already present on master. The
problem was that we only got a new connect configuration in the case
of WireGuard, which is not right. We should always get a new connect
configuration as '/disconnect' is called.