UI: Fix combobox width by recreating it every time
We create a new one every time because Gtk has some weird behaviour regarding the width of the combo box.
When we add items that are large, the combobox resizes to fit the content.
However, when we add items again that are all smaller (e.g. for a new server), the combo box does not shrink back.
The only proper way seems to be to recreate the combobox every time