UI: Show expiry labels according to spec and start con info earlier
See: https://docs.eduvpn.org/server/v3/client-implementation-notes.html

The global 'valid for' counter should not be shown if more than 24
hours are still left. The connection info now has a detailed `valid for` always
shown. Because we add a new row to the connection info we show the
Protocol info there instead of the title of the connection info