App: Add a configurable failover delay
This hopefully makes it so that the failover test is not started "too early".
It seems that on some systems, or just slow networks, the first ping/pong has a high chance of failing.
Adding a delay is hacky but at least this is configurable with an env variable so we can find a good middle ground
and the default of 1s is low enough that it should not be annoying.