Keyring: Improve detection and stop triggering it for every command
Fixes: #587

This makes sure that for the CLI the keyring is only used when
absolutely needed. The problem is that we cannot check if the keyring
is secure too often as that needs a test whether or not an item can be
added and removed to a collection.

For the GUI it's still used on startup (if the UI is opened) to show
the warning popup